Bugfix FXIOS-8602 #19093 ⁃ An unsightly animation occurs when the download toast message is updated (#25165)
* Organize download classes under one folder

* Add a delay to improve showing download completed toast

* Break down logic to handle success on didFinishDownloadingTo
